refactor(iwork): split the `TSWP` text layer into its own translation unit
`iwork_parser.cpp` held the container spines and the text reader they share.
Splitting them along Apple's own seam — `TSWP` storage, paragraph run tables
and the line separators inside a paragraph move to `iwork_text.cpp` — leaves
the parser the three per-app spines and the drawable geometry above them.

`Budget` moves with neither: the spines spend for slides and frames as the
text layer spends for paragraphs and runs, so it gets its own header.

This is the seam the module keeps splitting on as stages land — `TSD`
drawables, `TST` tiles, `TSS` styles each earn a file, and no split follows
app lines, because `iwork/` stays flat with one registry and one document for
all three apps. `PLAN.md` and `AGENTS.md` carry the argument: Apple factored
the format by framework, so a `.pages`, a `.key` and a `.numbers` share
everything below their spine — including the tile reader stage 6 writes for
Pages and stage 7 hands to Numbers.
